backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6
httpx==0.27.0

# Database
supabase==2.11.0
sqlalchemy==2.0.36
asyncpg==0.29.0

# Cache
redis==6.4.0

# AI/LLM
anthropic==0.43.0
openai==1.59.9
deepgram-sdk==3.8.0

# Data Processing
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Payments
stripe==11.4.1

# Email
sendgrid==6.11.0

# Monitoring
logfire==0.20.0

# PDF Generation
weasyprint==62.3
jinja2==3.1.3
matplotlib==3.8.2

# Utils
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
